Module discovery endpoint
