+++ backend/app/utils.py
@@ -1,5 +1,7 @@
 import os
+from typing import List
 from enum import StrEnum
+from pydantic import BaseModel
 
 class ENV_VARS(StrEnum):
     MQTT_BROKER_HOST = os.getenv('MQTT_BROKER_HOST', 'mqtt-broker')
@@ -8,17 +10,25 @@ class ENV_VARS(StrEnum):
     DB_PASS = os.getenv('DB_PASS', 'password')
     DB_NAME = os.getenv('DB_NAME', 'mydb')
 
-class Factors(StrEnum):
-    TEMPERATURE = "temp"
-    HUMIDITY = "humid"
-    NOISE = "noise"
-    LIGHT = "light"
-    PARTICULATE_MATTER = "pm"
-    CARBON_DIOXIDE = "co2"
-    OCCUPANCY = "occu"
+class SensorIn(BaseModel):
+    sensor_type: str
+    sensor_unit: str
 
-class Platforms:
-    TEST = "test"
-    PI = "pi"
-    ESP32 = "esp"
-    ARDUINO = "arduino"
+class ModuleIn(BaseModel):
+    room_id: int
+    x: int
+    y: int
+    z: int
+    sensors: List[SensorIn]
+
+class UnregModuleIn(BaseModel):
+    hw_id: str
+    room_id: int
+    x: int
+    y: int
+    z: int
+    sensors: List[SensorIn]
+
+class DiscoverableModule(BaseModel):
+    hw_id: str
+    sensor_count: int
